requirements.txt
git+https://github.com/ndgnuh/vietocr
python-doctr[torch]
python-doctr
streamlit
Pillow
numpy
networkx
